Splits exceptions in common + client
